confiance_de_trade_web/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
aiohttp==3.9.3
python-dotenv==1.0.1
PyYAML==6.0.1
pydantic==1.10.14
